improving auto format and creating util for excel
